requirements_cabo_crew.txt
# Core CrewAI and AI Framework Requirements
crewai==0.141.0
crewai-tools==0.51.1
langchain-openai==0.2.14
langchain-anthropic==0.2.8
langchain-google-genai==2.1.10
langchain-community==0.4.2
openai==1.95.1

# Data Processing and Structure
pydantic>=2.4.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# Optional: Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Optional: Web Interface
streamlit>=1.28.0

# Optional: Database Support
psycopg2-binary>=2.9.0  # PostgreSQL
sqlalchemy>=2.0.0

# Optional: Advanced NLP
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0